+++ crates/mdk-memory-storage/src/lib.rs
@@ -1507,6 +1507,7 @@ mod tests {
             image_key,
             image_nonce,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         nostr_storage.save_group(group.clone()).unwrap();
         let found_group = nostr_storage
@@ -1552,6 +1553,7 @@ mod tests {
             image_key,
             image_nonce,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         nostr_storage.save_group(group.clone()).unwrap();
         let relay_url1 = RelayUrl::parse("wss://relay1.example.com").unwrap();
@@ -1599,6 +1601,7 @@ mod tests {
             image_key,
             image_nonce,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         nostr_storage.save_group(group.clone()).unwrap();
         let group_exporter_secret_0 = GroupExporterSecret {
@@ -1751,6 +1754,7 @@ mod tests {
             image_key,
             image_nonce,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         nostr_storage.save_group(group.clone()).unwrap();
         let event_id = EventId::all_zeros();
@@ -1899,6 +1903,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         nostr_storage.save_group(group).unwrap();
 
@@ -1996,6 +2001,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         nostr_storage.save_group(group).unwrap();
 
@@ -2116,6 +2122,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         nostr_storage.save_group(group).unwrap();
 
@@ -2196,6 +2203,7 @@ mod tests {
             image_key,
             image_nonce,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
 
         // Save the group
@@ -2234,6 +2242,7 @@ mod tests {
             image_key,
             image_nonce,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
 
         // Save the group
@@ -2268,6 +2277,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(group.clone()).unwrap();
 
@@ -2324,6 +2334,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(group).unwrap();
 
@@ -2432,6 +2443,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(group).unwrap();
 
@@ -2469,6 +2481,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(group.clone()).unwrap();
 
@@ -2540,6 +2553,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(group).unwrap();
 
@@ -2592,6 +2606,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(group).unwrap();
 
@@ -2709,6 +2724,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(group1).unwrap();
 
@@ -2734,6 +2750,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(group2).unwrap();
 
@@ -2783,6 +2800,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(modified_group1).unwrap();
 
@@ -2838,6 +2856,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(group.clone()).unwrap();
 
@@ -2924,6 +2943,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
 
         let group2 = Group {
@@ -2941,6 +2961,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
 
         storage.save_group(group1.clone()).unwrap();
@@ -3034,6 +3055,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
 
         let group2 = Group {
@@ -3051,6 +3073,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
 
         storage.save_group(group1).unwrap();
@@ -3159,6 +3182,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
 
         let group2 = Group {
@@ -3176,6 +3200,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
 
         storage.save_group(group1).unwrap();
@@ -3285,6 +3310,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
 
         storage.save_group(group).unwrap();
@@ -3341,6 +3367,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(group).unwrap();
 
@@ -3598,6 +3625,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(group).unwrap();
 
@@ -3710,6 +3738,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(group).unwrap();
 
@@ -3783,6 +3812,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(group).unwrap();
 
@@ -3868,6 +3898,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(group).unwrap();
 
@@ -3972,6 +4003,7 @@ mod tests {
                 image_key: None,
                 image_nonce: None,
                 self_update_state: SelfUpdateState::Required,
+                disappearing_message_secs: None,
             }
         }