+++ crates/mdk-sqlite-storage/src/messages.rs
@@ -389,6 +389,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
 
         // Save the group
@@ -502,6 +503,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(group).unwrap();
 
@@ -565,6 +567,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
 
         let group_2 = Group {
@@ -582,6 +585,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
 
         storage.save_group(group_1).unwrap();
@@ -791,6 +795,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(group).unwrap();
 
@@ -879,6 +884,7 @@ mod tests {
             image_key: None,
             image_nonce: None,
             self_update_state: SelfUpdateState::Required,
+            disappearing_message_secs: None,
         };
         storage.save_group(group).unwrap();
         group_id